requirements.txt
# requirements.txt

# 核心套件
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# 視覺化
plotly>=5.17.0

# 技術分析
ta>=0.11.0
yfinance>=0.2.28

# API 與工具
requests>=2.31.0
python-dotenv>=1.0.0

# 資料庫
sqlite3  # Python 內建，無需安裝

# 日期時間處理
python-dateutil>=2.8.2

# HTTP 請求
urllib3>=2.0.0
